SecureAgent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "SecureAgent"
version = "0.1.3"
description = "Agentic Identity Framework abstracting Keycloak/OIDC complexity for AI Agents"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["security", "oidc", "keycloak", "agents", "ai"]
authors = [
  {name = "Sam Armstrong", email = "DO_NOT_EMAIL_USE_ISSUES@github.com"},
]
dependencies = [
  "fastapi",
  "python-keycloak",
  "httpx",
  "pydantic",
  "click>=8.0",
]

[project.optional-dependencies]
test = [
  "pytest",
  "pytest-mock",
  "pytest-cov",
  "pytest-asyncio",
  "python-keycloak", 
  "httpx"
]

[project.scripts]
secureagent = "SecureAgent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/SecureAgent"]
